fix(review): move shared fixtures/helpers to conftest, fix guard, remove box-comments
- Move deep_workflow_section fixture to tests/skills/conftest.py using the
  correct for-loop implementation (fixes the always-returning while loop in
  the design intent file)
- Move _extract_step_section helper to conftest and import in both test files
  to eliminate verbatim duplication
- Add assert historical_pos != -1 guard before ordering assertion so missing
  '## Historical Context' produces a clear diagnostic
- Remove five decorative box-comment section dividers (REQ-INTENT-001/002,
  Regression Guards, Subagent Templates) — REQ IDs already appear in docstrings
